+++ frontend/public/config/config.json
@@ -1,5 +1,5 @@
 {
-  "CONTRACT_ADDRESS": "0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0",
+  "CONTRACT_ADDRESS": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
   "SCAN_LINK": "http://127.0.0.1:8545/token/0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512",
   "NETWORK": {
     "NAME": "Localhost8545",
@@ -15,5 +15,9 @@
   "MARKETPLACE": "Opeansea",
   "MARKETPLACE_LINK": "https://opensea.io/collection/holland-lop-gene",
   "SHOW_BACKGROUND": false,
-  "PHASE": "WLSale"
+  "PHASE": "WLSale",
+  "WHITELIST": [
+    "0x72A1241a02b45c63cd2731A9136Fe44FcD781AA1",
+    "0x70997970C51812dc3A010C7d01b50e0d17dc79C8"
+  ]
 }
